Add library(fcuk) to the .Rprofile file

After calling this function, fcuk will be launched everytime you launch your current R project (or any project if you use option global = TRUE).

add_fcuk_to_rprofile(global = TRUE)

Arguments

global
Whether to add it to the global .Rprofile (TRUE) or just to the .Rprofile file of your current project (FALSE, the default).

Examples

fcuk::add_fcuk_to_rprofile()

#> ⚠ fcuk was already in ~/.Rprofile